+++ tests/testthat/test-idxFinalT.R
@@ -20,3 +20,15 @@ test_that("N, NResource and getTimes expose stored arrays and times", {
     expect_identical(NResource(sim), sim@n_pp)
     expect_identical(getTimes(sim), c(0, 0.5, 1))
 })
+
+test_that("MizerSim accessors are registered as S3 methods", {
+    expect_identical(utils::getS3method("validSim", "MizerSim"), validSim.MizerSim)
+    expect_identical(utils::getS3method("N", "MizerSim"), N.MizerSim)
+    expect_identical(utils::getS3method("NResource", "MizerSim"), NResource.MizerSim)
+    expect_identical(utils::getS3method("finalN", "MizerSim"), finalN.MizerSim)
+    expect_identical(utils::getS3method("finalNResource", "MizerSim"), finalNResource.MizerSim)
+    expect_identical(utils::getS3method("idxFinalT", "MizerSim"), idxFinalT.MizerSim)
+    expect_identical(utils::getS3method("getTimes", "MizerSim"), getTimes.MizerSim)
+    expect_identical(utils::getS3method("getEffort", "MizerSim"), getEffort.MizerSim)
+    expect_identical(utils::getS3method("getParams", "MizerSim"), getParams.MizerSim)
+})